+++ R/add_n.R
@@ -4,10 +4,23 @@
 #' total number of non-missing (or missing) observations
 #'
 #' @param x Object with class `tbl_summary` from the [tbl_summary] function
-#' @param missing Logical argument indicating whether to print N (`missing = FALSE`),
-#' or N missing (`missing = TRUE`).  Default is `FALSE`
+#' @param statistic String indicating the statistic to report. Default is the
+#' number of non-missing observation for each variable, `statistic = "{n}"`.
+#' Other statistics available to report include:
+#' * `"{N}"` total number of observations,
+#' * `"{n}"` number of non-missing observations,
+#' * `"{n_miss}"` number of missing observations,
+#' * `"{p}"` percent non-missing data,
+#' * `"{p_miss}"` percent missing data
+#' The argument uses [glue::glue] syntax and multiple statistics may be reported,
+#' e.g. `statistic = "{n} / {N} ({p}%)"`
+#' @param col_label String indicating the column label.  Default is `"**N**"`
+#' @param footnote Logical argument indicating whether to print a footnote
+#' clarifying the statistics presented. Default is `FALSE`
 #' @param last Logical indicator to include N column last in table.
 #' Default is `FALSE`, which will display N column first.
+#' @param missing DEPRECATED. Logical argument indicating whether to print N
+#' (`missing = FALSE`), or N missing (`missing = TRUE`).  Default is `FALSE`
 #' @family tbl_summary tools
 #' @author Daniel D. Sjoberg
 #' @export
@@ -21,29 +34,47 @@
 #' @section Example Output:
 #' \if{html}{\figure{tbl_n_ex.png}{options: width=50\%}}
 
-add_n <- function(x, missing = FALSE, last = FALSE) {
+add_n <- function(x, statistic = "{n}", col_label = "**N**", footnote = FALSE,
+                  last = FALSE, missing = NULL) {
   # checking that input is class tbl_summary
   if (class(x) != "tbl_summary") stop("x must be class 'tbl_summary'")
 
-  # counting non-missing N (or missing N)
+  # defining function to round percentages -------------------------------------
+  percent_fun <- getOption("gtsummary.tbl_summary.percent_fun",
+                           default = style_percent)
+
+  # DEPRECATED specifying statistic via missing argument -----------------------
+  if (!is.null(missing)) {
+    lifecycle::deprecate_warn("1.2.2",
+                              "gtsummary::add_n(missing = )",
+                              "gtsummary::add_n(statistic = )")
+    if (identical(missing, TRUE)) {
+      statistic = "{n_miss}"
+      col_label = "**N Missing**"
+    }
+  }
+
+  # counting non-missing N (or missing N) --------------------------------------
   counts <-
     x$meta_data %>%
     select(c("variable")) %>%
     mutate(
       row_type = "label",
-      n_var = map_chr(
-        .data$variable,
-        ~ case_when(
-          missing == FALSE ~ sum(!is.na(x$inputs$data[[.x]])),
-          missing == TRUE ~ sum(is.na(x$inputs$data[[.x]]))
-        )
-      )
+      N = nrow(x$inputs$data),
+      n = purrr::map_int(.data$variable, ~ sum(!is.na(x$inputs$data[[.x]]))),
+      n_miss = purrr::map_int(.data$variable, ~ sum(is.na(x$inputs$data[[.x]]))),
+      p = percent_fun(.data$n / .data$N),
+      p_miss = percent_fun(.data$n_miss / .data$N),
+      statistic = glue(statistic) %>% as.character()
     ) %>%
-    set_names(c(
-      "variable", "row_type", ifelse(missing == FALSE, "n", "n_missing")
-    ))
+    select(.data$variable, .data$row_type, n = .data$statistic)
 
-  # merging result with existing tbl_summary
+  # DEPRECATED specifying column name via `missing` argument -------------------
+  if (identical(missing, TRUE)) {
+    counts <- rename(counts, n_missing = .data$n)
+  }
+
+  # merging result with existing tbl_summary -----------------------------------
   if (last == FALSE) {
     table_body <-
       x$table_body %>%
@@ -57,7 +88,7 @@ add_n <- function(x, missing = FALSE, last = FALSE) {
       left_join(counts, by = c("variable", "row_type"))
   }
 
-  # replacing old table_body with new
+  # replacing old table_body with new ------------------------------------------
   x$table_body <- table_body
 
   x$table_header <-
@@ -65,12 +96,29 @@ add_n <- function(x, missing = FALSE, last = FALSE) {
     left_join(x$table_header, by = "column") %>%
     table_header_fill_missing()
 
-  # updating header
-  if (missing == FALSE) {
-    x <- modify_header_internal(x, n = "**N**")
+  # Adding footnote if requested -----------------------------------------------
+  if (footnote == TRUE) {
+    x$table_header <-
+      x$table_header %>%
+      mutate(
+        footnote= map2(
+          .data$column, .data$footnote,
+          function(x1, y1) {
+            if (x1 %in% c("n"))
+              return(c(y1, paste("Statistics presented:", stat_to_label(statistic))))
+            return(y1)
+          }
+        )
+      )
   }
-  else if (missing == TRUE) {
-    x <- modify_header_internal(x, n_missing = "**N Missing**")
+
+  # updating header ------------------------------------------------------------
+  # DEPRECATED specifying column name via `missing` argument -------------------
+  if (identical(missing, TRUE)) {
+    x <- modify_header_internal(x, n_missing = col_label)
+  }
+  else {
+    x <- modify_header_internal(x, n = col_label)
   }
 
   # updating gt and kable calls with data from table_header
@@ -81,4 +129,16 @@ add_n <- function(x, missing = FALSE, last = FALSE) {
 
   # returning tbl_summary object
   return(x)
+}
+
+stat_to_label <- function(x) {
+  x <- stringr::str_replace_all(x, fixed("{N}"), fixed("Total N"))
+  x <- stringr::str_replace_all(x, fixed("{n}"), fixed("N non-Missing"))
+  x <- stringr::str_replace_all(x, fixed("{n_miss}"), fixed("N Missing"))
+  x <- stringr::str_replace_all(x, fixed("{p}%"), fixed("% non-Missing"))
+  x <- stringr::str_replace_all(x, fixed("{p}"), fixed("% non-Missing"))
+  x <- stringr::str_replace_all(x, fixed("{p_miss}%"), fixed("% Missing"))
+  x <- stringr::str_replace_all(x, fixed("{p_miss}"), fixed("% Missing"))
+
+  x
 }