+++ R/tbl_summary.R
@@ -45,8 +45,7 @@
 #' e.g. `sort = list(everything() ~ "frequency")`
 #' @param percent Indicates the type of percentage to return. Must be one of
 #' `"column"`, `"row"`, or `"cell"`. Default is `"column"`.
-#' @param group DEPRECATED. Previously required to work with the 'group'
-#' argument in [add_p]. No longer required.
+#' @param group DEPRECATED. Migrated to [add_p]
 #'
 #' @section select helpers:
 #' \href{http://www.danieldsjoberg.com/gtsummary/articles/tbl_summary.html#select_helpers}{Select helpers}
@@ -121,6 +120,19 @@
 #'     statistic = list(all_continuous() ~ "{mean} ({sd})"),
 #'     digits = list(vars(age) ~ c(0, 1))
 #'   )
+#'
+#' # for convenience, you can also pass named lists to any arguments
+#' # that accept formulas (e.g label, digits, etc.)
+#' trial %>%
+#'    dplyr::select(age, grade, response, trt) %>%
+#'    tbl_summary(
+#'     by = trt,
+#'     label = list(age = "Patient Age"),
+#'     statistic = list(all_continuous() ~ "{mean} ({sd})"),
+#'     digits = list(vars(age) ~ c(0, 1))
+#'   )
+#'
+#'
 #' @section Example Output:
 #' \if{html}{Example 1}
 #'
@@ -131,10 +143,10 @@
 #' \if{html}{\figure{tbl_summary_ex2.png}{options: width=45\%}}
 
 tbl_summary <- function(data, by = NULL, label = NULL, statistic = NULL,
-                        digits = NULL, type = NULL, value = NULL, group = NULL,
+                        digits = NULL, type = NULL, value = NULL,
                         missing = c("ifany", "always", "no"),
                         missing_text = "Unknown", sort = NULL,
-                        percent = c("column", "row", "cell")) {
+                        percent = c("column", "row", "cell"),  group = NULL) {
 
   # converting bare arguments to string ----------------------------------------
   by <- enquo_to_string(rlang::enquo(by), arg_name = "by")
@@ -159,27 +171,44 @@ tbl_summary_ <- function(data, by = NULL, label = NULL, statistic = NULL,
                         missing = c("ifany", "always", "no"),
                         missing_text = "Unknown", sort = NULL,
                         percent = c("column", "row", "cell"), group = NULL) {
-  # matching arguments
+  # matching arguments ---------------------------------------------------------
   missing <- match.arg(missing)
   percent <- match.arg(percent)
 
-  # ungrouping data
+  # ungrouping data ------------------------------------------------------------
   data <- data %>% ungroup()
 
-  # deprecation note about group
+  # deleting obs with missing by values ----------------------------------------
+  # saving variable labels
+  if (!is.null(by) && sum(is.na(data[[by]])) > 0) {
+    message(glue(
+      "{sum(is.na(data[[by]]))} observations missing `{by}` have been removed. ",
+      "To include these observations, use `forcats::fct_explicit_na()` on `{by}` ",
+      "column before passing to `tbl_summary()`."
+    ))
+    lbls <- purrr::map(data, ~attr(.x, "label"))
+    data <- data[!is.na(data[[by]]), ]
+
+    # re-applying labels---I think this will NOT be necessary after dplyr 0.9.0
+    for (i in names(lbls)) {
+      attr(data[[i]], "label") <- lbls[[i]]
+    }
+  }
+
+  # deprecation note about group -----------------------------------------------
   if (!rlang::quo_is_null(rlang::enquo(group))) {
-    stop_defunct(glue(
+    stop(glue(
       "Passing the 'group' argument in 'tbl_summary()' is defunct.\n",
       "Please pass the column in 'add_p()'. For example,\n\n",
       "tbl_summary() %>% add_p(group = varname)"
     ))
   }
 
-  # will return call, and all object passed to in tbl_summary call
+  # will return call, and all object passed to in tbl_summary call -------------
   # the object func_inputs is a list of every object passed to the function
   tbl_summary_inputs <- as.list(environment())
 
-  # removing variables with unsupported variable types from data
+  # removing variables with unsupported variable types from data ---------------
   classes_expected <- c("character", "factor", "numeric", "logical", "integer")
   var_to_remove <-
     map_lgl(data, ~ class(.x) %in% classes_expected %>% any()) %>%
@@ -196,17 +225,17 @@ tbl_summary_ <- function(data, by = NULL, label = NULL, statistic = NULL,
     ))
   }
 
-  # checking function inputs
+  # checking function inputs ---------------------------------------------------
   tbl_summary_input_checks(
     data, by, label, type, value, statistic,
     digits, missing, missing_text, sort
   )
 
-  # converting tidyselect formula lists to named lists
+  # converting tidyselect formula lists to named lists -------------------------
   type <- tidyselect_to_list(data, type, input_type = "type")
   value <- tidyselect_to_list(data, value, input_type = "value")
 
-  # creating a table with meta data about each variable
+  # creating a table with meta data about each variable ------------------------
   meta_data <- tibble(
     variable = names(data),
     # assigning class, if entire var is NA, then assigning class NA
@@ -218,13 +247,13 @@ tbl_summary_ <- function(data, by = NULL, label = NULL, statistic = NULL,
   # excluding by variable
   if (!is.null(by)) meta_data <- meta_data %>% filter(!!parse_expr("variable != by"))
 
-  # converting tidyselect formula lists to named lists
+  # converting tidyselect formula lists to named lists -------------------------
   label <- tidyselect_to_list(data, label, .meta_data = meta_data, input_type = "label")
   statistic <- tidyselect_to_list(data, statistic, .meta_data = meta_data, input_type = "statistic")
   digits <- tidyselect_to_list(data, digits, .meta_data = meta_data, input_type = "digits")
   sort <- tidyselect_to_list(data, sort, .meta_data = meta_data)
 
-  # assigning variable characteristics
+  # assigning variable characteristics -----------------------------------------
   meta_data <-
     meta_data %>%
     mutate(
@@ -238,7 +267,7 @@ tbl_summary_ <- function(data, by = NULL, label = NULL, statistic = NULL,
       sort = assign_sort(.data$variable, .data$summary_type, sort)
     )
 
-  # calculating summary statistics
+  # calculating summary statistics ---------------------------------------------
   table_body <-
     meta_data %>%
     mutate(
@@ -262,7 +291,7 @@ tbl_summary_ <- function(data, by = NULL, label = NULL, statistic = NULL,
     select(c("variable", "summary_type", "stat_table")) %>%
     unnest(!!sym("stat_table"))
 
-  # table of column headers
+  # table of column headers ----------------------------------------------------
   table_header <-
     tibble(column = names(table_body) %>% setdiff("summary_type")) %>%
     table_header_fill_missing() %>%
@@ -275,7 +304,7 @@ tbl_summary_ <- function(data, by = NULL, label = NULL, statistic = NULL,
       )
     )
 
-  # returning all results in a list
+  # returning all results in a list --------------------------------------------
   results <- list(
     gt_calls = eval(gt_tbl_summary),
     kable_calls = eval(kable_tbl_summary),